JGM_RAG_Chatbot/requirements.txt
# ===== CORE DEPENDENCIES =====
flask==3.0.0
python-dotenv==1.0.0

# ===== GOOGLE AI =====
google-adk>=0.1.0

# ===== DATA PROCESSING =====
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0  # For Excel files

# ===== TEXT PROCESSING =====
rapidfuzz>=3.0.0
scikit-learn>=1.3.0

# ===== VISUALIZATION =====
matplotlib>=3.7.0
folium>=0.14.0

# ===== AI/ML (OPTIONAL - for Ollama fallback) =====
llama-index>=0.9.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-ollama>=0.1.0

# ===== PRODUCTION =====
gunicorn>=21.2.0  # For production deployment
waitress>=2.1.2   # Alternative WSGI server (Windows compatible)

# ===== DEVELOPMENT =====
pytest>=7.4.0     # For testing
black>=23.0.0     # Code formatting